write-behind/write-pubsub/gae/requirements.txt
google-cloud-pubsub==2.13.9
Flask==2.1.0
redis==4.1.1
gunicorn==20.1.0
